requirements.txt
fastapi
uvicorn[standard]
streamlit
httpx
python-dotenv
loguru
chromadb
pypdf